fix: round-1 max-review findings — 3 shm lifecycle races + 9 defects
Fixes all 12 actionable findings from convergence round 1 (44 verified ->
15 distinct; the engine-dedup finding gets its own commit; the anchored
scope gap is a round-2 process fix):

shm lifecycle races (each with poke/fork regression tests):
- spmc_release_claim rollback is generation-conditional — an attach
  rollback racing a create_shm re-init leaks its claim instead of
  clobbering the NEW ring's slot (leak-not-clobber).
- spmc_reset_table arms the seqlock generation (create's odd/even
  protocol) and shm_subscribe on both gating rings snapshots + re-checks
  it — recover_shm can no longer silently erase a concurrent attach.
- force_detach_consumer is epoch-guarded: compare-and-retire
  ACTIVE@e -> RETIRED@e+1 with InvalidInput on mismatch; graceful detach
  slot reuse can no longer be retired out from under a healthy consumer.
  Consumers expose shm_slot_epoch(); guide updated.

Other fixes:
- lib.rs: spmc/spmc_bytes cfg-gated on target_has_atomic="64"
  (portability regression — AtomicU64 registry bitmaps).
- broadcast_bytes/anchored/anchored_bytes: drained check is <= (a dead
  producer's stale latest > tail no longer hangs the drain; regression
  test livelocks without it); the prescribed latest.min(tail) clamp was
  itself wrong and is documented-rejected (latest legitimately leads
  tail; availability blocks reads past tail).
- broadcast/anchored: skip_to_latest saturates and never moves backwards.
- spmc/spmc_bytes from_shm: one real registry rescan at construction —
  attached producers' len/is_full no longer lie until the first push
  (documented livelock).
- create_spmc_region rejects max_consumers > u32::MAX (header
  truncation); SPSC/SPMC creates zero the bcast stride word and
  OFF_MAX_CONSUMERS on reused fds (promised-zero sweep); bcast element
  stride validated by exact equality via a header word at offset 152
  (lower-alignment T attach now rejected).
- pop_into clears the caller's buffer on entry (doc now true on every
  error path).
- bench_broadcast: f64 throughput math (32-bit overflow).
- spmc_bytes + anchored_bytes: span-valued starving flag — the flag
  carries the blocked push's actual pad+record span, making the lag
  filter real (the old worst-case threshold degenerated to ALIGN and
  filtered nobody); exact-per-episode + conservative-under-staleness
  arguments documented; bd rust-rb-fzv closed.

Gates: 264/317 tests; full Miri across all six ring suites (151 tests,
exit 0); fmt/clippy/doc -D warnings both feature sets; MSRV 1.70;
volatile-cfg suite green.

+++ examples/bench_broadcast.rs
@@ -151,7 +151,9 @@ fn run_copy_stream<const L: usize>(iters: u64, capacity: usize, cores: &[usize])
     let stats = consumer.join().unwrap();
 
     let ns_per_op = elapsed.as_nanos() as f64 / iters as f64;
-    let gb_per_s = (iters as usize * L) as f64 / elapsed.as_nanos() as f64;
+    // Multiply in f64: `iters as usize * L` truncates the u64 count and can
+    // overflow on 32-bit targets (bytes = iters * L easily exceeds 2^32).
+    let gb_per_s = iters as f64 * L as f64 / elapsed.as_nanos() as f64;
     println!(
         "BCAST_stream {L:>3}B      {ns_per_op:>6.2} ns/push  {gb_per_s:>6.2} GB/s   {:>5} ms",
         elapsed.as_millis()
@@ -186,7 +188,8 @@ fn run_copy_pingpong<const L: usize>(rounds: usize, capacity: usize, cores: &[us
         pop_ns += t.elapsed().as_nanos();
     }
     std::hint::black_box(sink);
-    let msgs = (rounds * capacity) as f64;
+    // Multiply in f64 (message count can exceed a 32-bit usize).
+    let msgs = rounds as f64 * capacity as f64;
     let pop_gb_per_s = msgs * L as f64 / pop_ns as f64;
     println!(
         "BCAST_pingpong {L:>3}B    push {:>6.2} ns  pop {:>6.2} ns  (pop copy {pop_gb_per_s:>6.2} GB/s)",
